+++ server/modules/monitor/account_mon.h
@@ -61,4 +61,4 @@ typedef struct {
 #define MONITOR_INTERVAL 10000 // in milliseconds
 #define MONITOR_DEFAULT_ID 1UL // unsigned long value
 
-uint32_t account_monitor_find_shard(ACCOUNT_MONITOR *, uint32_t);
+uintptr_t account_monitor_find_shard(ACCOUNT_MONITOR *, uintptr_t);
